Rename cprint.rs to cprint_macros.rs
