+++ src/lib.rs
@@ -1,4 +1,4 @@
 mod color;
-mod cprint;
+mod cprint_macros;
 
 pub use color::{Color, Coloration};
